Add files via upload
"Exploration of Bistability and Stochastic Influence in the Toggle Switch Gene Regulatory Circuit"

This repository contains code for analyzing the bistable behaviour in a toggle switch gene regulatory circuit under stochastic influences. When run, the code generates 14 plots detailed in the document "HCKZ5-CELL0014-REPORT". The code was developed and tested in JupyterLab version 4.0.7 with Python 3.8.18.